api/requirements.txt
fastapi
uvicorn[standard]
pandas
matplotlib
seaborn
sqlalchemy
langchain
langchain-community
langchain-core
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-sdk
langchain-google-genai
pydantic
google-generativeai